* Added serialization ID for TaxonomyNode to guarantee compatibility of MPA project files between different versions. * Removed classes which are not needed anymore. * Removed old tests for graph database. * Fixed some TODOs.

+++ trunk/src/de/mpa/client/settings/CruxParameters.java
@@ -50,13 +50,18 @@ public class CruxParameters extends ParameterMap {
 	@Override
 	public File toFile(String path) throws IOException {
 		// Append extension if it's missing
-		// TODO: add check for whether path does point to a file (and not a directory)
+		
 		if (!path.endsWith(".params")) {
 			path += ".params";
 		}
 		
 		// Create new file at specified path
 		File file = new File(path);
+		
+		// Check for whether path does point to a file (and not a directory)
+		if(!file.isFile()) {
+			throw new IOException();
+		}
 		
 		// Set up writer for parameter file
 		BufferedWriter bw = new BufferedWriter(new FileWriter(file));
